requirements.txt
Flask
python-dotenv
requests
google-generativeai